+++ include/HX2.hpp
@@ -3,6 +3,6 @@
 
 #include "parameters.hpp"
 
-void HX2(double y_hx2[length_hx], double Ts_HX2_L, int step, Parameters &params);
+void HX2(float y_hx2[length_hx], float Ts_HX2_L, int step, Parameters &params);
 
 #endif // HX2_HPP
