+++ include/csr_matrix.hpp
@@ -2,6 +2,6 @@
 #define CSR_MATRIX_HPP
 #include "parameters.hpp"
 
-void sparse_csr(double dense[N][N], CSRMatrix& csr);
+void sparse_csr(float dense[N][N], CSRMatrix& csr);
 
 #endif // CSR_MATRIX_HPP
