+++ include/data_saving.hpp
@@ -5,17 +5,17 @@
 #include <string>
 #include "parameters.hpp"
 
-void save_results(const double rho_matrix[time_span], 
-                  const double phi_middle_matrix[time_span * 2],
-                  const double ci_middle_matrix[time_span * 6], 
-                  const double temperature_fuel_middle_matrix[time_span],
+void save_results(const float rho_matrix[time_span], 
+                  const float phi_middle_matrix[time_span * 2],
+                  const float ci_middle_matrix[time_span * 6], 
+                  const float temperature_fuel_middle_matrix[time_span],
                   const std::string& folder);
 
-void save_spacial_results(const double rho[N], 
-                          const double y_n[length_neutr], 
-                          const double y_th[length_th], 
-                          const double y_hx1[length_hx], 
-                          const double y_hx2[length_hx], 
+void save_spacial_results(const float rho[N], 
+                          const float y_n[length_neutr], 
+                          const float y_th[length_th], 
+                          const float y_hx1[length_hx], 
+                          const float y_hx2[length_hx], 
                           const std::string& folder);
 
 #endif // DATA_SAVING_HPP
