+++ include/matrix_LU.hpp
@@ -2,7 +2,7 @@
 #define MATRIX_INVERSE_HPP
 #include "parameters.hpp"
 
-void invert_LU(double A[N][N]);
+void invert_LU(float A[N][N]);
 // void inverse_sparse(const CSRMatrix &A_csr, CSRMatrix &A_inv) ;
 
 #endif // MATRIX_INVERSE_HPP
