+++ include/thermal_hydraulics.hpp
@@ -4,6 +4,6 @@
 #include "parameters.hpp"
 
 // Function to solve the thermal hydraulics problem
-void thermal_hydraulics(double y_th[length_th], const double q_prime[N], double Ts_core_0, int step, Parameters &params);
+void thermal_hydraulics(float y_th[length_th], const float q_prime[N], float Ts_core_0, int step, Parameters &params);
 
 #endif // THERMAL_HYDRAULICS_HPP
